+++ src/Products/ZCTextIndex/OkapiIndex.py
@@ -19,7 +19,7 @@
 
 from BTrees.IIBTree import IIBucket
 from BTrees.Length import Length
-from zope.interface import implements
+from zope.interface import implementer
 
 from Products.ZCTextIndex.interfaces import IIndex
 from Products.ZCTextIndex.BaseIndex import BaseIndex
@@ -27,9 +27,8 @@ from Products.ZCTextIndex.BaseIndex import inverse_doc_frequency
 from Products.ZCTextIndex.BaseIndex import scaled_int
 
 
+@implementer(IIndex)
 class OkapiIndex(BaseIndex):
-
-    implements(IIndex)
 
     # BM25 free parameters.
     K1 = 1.2